Do not reverse the order of files in Query Directory calls
Reversing the order of the files resulted in a vary esoteric bug that
broke some SKSE plugins.  The order of the files is not guaranteed
by the Windows API but USVFS should try to maintain the order provided
by Windows to promote compatibility.

This change probably results in reduced performance.